# docs/submission/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SAHYOG Blockchain Intelligence",
    "description": "Complaint intake, replayable ingestion, tracing, attribution and VASP federation.",
    "version": "0.16.0"
  },
  "paths": {
...
    "/api/v1/integrations/sahyog/webhook": {
      "post": {
        "tags": [
          "SAHYOG Integration"
        ],
        "summary": "Webhook",
        "operationId": "webhook_api_v1_integrations_sahyog_webhook_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Complaint"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/integrations/sahyog/complaints/{ref}/import": {
      "post": {
        "tags": [
          "SAHYOG Integration"
        ],
        "summary": "Import Complaint",
        "operationId": "import_complaint_api_v1_integrations_sahyog_complaints__ref__import_post",
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ref"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AttachmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "extracted_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extracted Json"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sha256",
          "extracted_json",
          "created_at"
        ],
        "title": "AttachmentOut"
      },
...
      "CaseOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "case_ref": {
            "type": "string",
            "title": "Case Ref"
          },
          "sahyog_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sahyog Ref"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "gang_case_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gang Case Id"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionOut"
            },
            "type": "array",
            "title": "Transactions"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentOut"
            },
            "type": "array",
            "title": "Attachments"
          }
        },
        "type": "object",
        "required": [
          "id",
          "case_ref",
          "sahyog_ref",
          "title",
          "status",
          "gang_case_id",
          "scenario_id",
          "created_by",
          "created_at",
          "source",
          "transactions",
          "attachments"
        ],
        "title": "CaseOut"
      },
...
      "Chain": {
        "type": "string",
        "enum": [
          "ethereum",
          "bnb",
          "polygon",
          "tron",
          "bitcoin"
        ],
        "title": "Chain"
      },
      "Complaint": {
        "properties": {
          "ref": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_/-]{3,80}$",
            "title": "Ref"
          },
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 3,
            "title": "Title"
          },
          "transaction": {
            "$ref": "#/components/schemas/TransactionInput"
          },
          "source": {
            "type": "string",
            "enum": [
              "simulated",
              "live"
            ],
            "title": "Source",
            "default": "simulated"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ref",
          "title",
          "transaction"
        ],
        "title": "Complaint"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TransactionInput": {
        "properties": {
          "chain": {
            "$ref": "#/components/schemas/Chain"
          },
          "tx_hash": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 66
              },
              {
                "type": "null"
              }
            ],
            "title": "Tx Hash"
          },
          "victim_address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Victim Address"
          },
          "suspect_address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Suspect Address"
          },
          "token": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Z0-9._-]{1,20}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          },
          "amount": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9]{1,38}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "decimals": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 30.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Decimals"
          },
          "tx_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tx Time"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "chain"
        ],
        "title": "TransactionInput"
      },
      "TransactionOut": {
        "properties": {
          "chain": {
            "$ref": "#/components/schemas/Chain"
          },
          "tx_hash": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 66
              },
              {
                "type": "null"
              }
            ],
            "title": "Tx Hash"
          },
          "victim_address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Victim Address"
          },
          "suspect_address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Suspect Address"
          },
          "token": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Z0-9._-]{1,20}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          },
          "amount": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9]{1,38}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "decimals": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 30.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Decimals"
          },
          "tx_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tx Time"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "chain",
          "id",
          "status"
        ],
        "title": "TransactionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}